modify .gitignore
