+++ .gitignore
@@ -31,3 +31,7 @@ build/
 
 ### VS Code ###
 .vscode/
+
+### properties
+application-dev.properties
+application-prod.properties
