--- src/main/java/com/nhnacademy/minidooraytaskapi/getTag/controller/GetTagController.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.minidooraytaskapi.getTag.controller;
-
-public class GetTagController {
-}
