--- src/main/java/com/nhnacademy/minidooraytaskapi/getTag/dto/GetTagDto.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.minidooraytaskapi.getTag.dto;
-
-public interface GetTagDto {
-}
