--- src/main/java/com/nhnacademy/minidooraytaskapi/getTag/repository/GetTagRepository.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.minidooraytaskapi.getTag.repository;
-
-public class GetTagRepository {
-}
