--- src/main/java/com/nhnacademy/minidooraytaskapi/getTag/service/DefaultGetTagService.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.minidooraytaskapi.getTag.service;
-
-public class DefaultGetTagService {
-}
