--- src/main/java/com/nhnacademy/minidooraytaskapi/getTag/service/GetTagService.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.minidooraytaskapi.getTag.service;
-
-public interface GetTagService {
-}
