+++ src/main/java/com/nhnacademy/minidooraytaskapi/get_tag/controller/GetTagController.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.minidooraytaskapi.get_tag.controller;
+
+public class GetTagController {
+}
