+++ src/main/java/com/nhnacademy/minidooraytaskapi/get_tag/dto/GetTagDto.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.minidooraytaskapi.get_tag.dto;
+
+public interface GetTagDto {
+}
