+++ src/main/java/com/nhnacademy/minidooraytaskapi/get_tag/repository/GetTagRepository.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.minidooraytaskapi.get_tag.repository;
+
+public class GetTagRepository {
+}
