+++ src/main/java/com/nhnacademy/minidooraytaskapi/get_tag/service/DefaultGetTagService.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.minidooraytaskapi.get_tag.service;
+
+public class DefaultGetTagService {
+}
