+++ src/main/java/com/nhnacademy/minidooraytaskapi/get_tag/service/GetTagService.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.minidooraytaskapi.get_tag.service;
+
+public interface GetTagService {
+}
